Workaround some problems on selecting data with keyboard on a ComboEntry.
For some reason, data is not set when typing or pressing enter on the
completion. It only works with the mouse (both clicking on completion and
scrolling results).

The workaround is a very ugly hack and we should remove it as soon as we have
the knowledge of what is happening.