Add a cell-edigint-started signal to objectlist
A cell-editing-started object will be emitted when the user starts to edit a
cell, that is, a cell that contains a spin button, a combobox, etc.
That's useful for example, when we have a column with a spin, and we want to
change it's upper/lower property depending on the row. There must be other
use cases too.