Make reversed column sorting work again.
results.reversed() is a sqlobject specific api. Since we are using storm on
Stoq, this doesn't work anymore. Also, Storm doesn't have any api like that.

Instead of relying on an specific orm api, reverse the list on load time.
This doesn't change anything since enumerate would iterate over it, we are just
creating it before and reversing it for enumerate. Also it makes
lazyobjectlist's results less api specific.