Make sure load_items_from_results will load the same amount of items
Lets say load_items_from_results received 10 as start and 20 as the end. If
results till 19 were already loaded, it would make it just load 19-20 (1 item).
After this, if the same happened, start would be moved to 19, and end to 29,
making it load the same amount it would if it didn't have any cache.

r=romaia