Propagate changes on an invalid widget after it become valid
The problem were: We have a widget in an invalid state. In that state, for
obvious reasons, the content in it is not propagated to the model. Something
changes and the widget becomes valid. But after that, the content (that now is
valid) wasn't being propagated to the model.

This fixes that problem (the content will be propagated to model as soon as the
widget becomes valid). Also, there's a new test that would break if this fix
wasn't here.

r=romaia