Changes API for selecting files.
Now, open() (renamed to selectfile(), see below) creates and returns a
gtk.FileChooserDialog so the user app can edit and adapt its behavior.
The old API is still supported for now, but could be removed without
further notice.

This patch also removes use of built-in names of Python, in this case
open() and filter(): they are reserved words and should not be used to
define functions or parameters.

Change-Id: If2cdfef0228dac579de9ff109995dcfba2b4ee0e