Allow conversion to str/unicode for widgets in proxy
The conversion to str/unicode is trivial, so if we have a widget on a proxy
with data_type of str and pass an int/decimal/datetime/etc to it, it should
convert instead of raising an exception. For all other cases, where the
conversion will never be as trivial as this, the old behavior is still true.

Change-Id: Ibdffe2fd5492c3f70b59a5eeff0810920a115b83