Fix _update_current_object for ProxyEntry
We've been just using ENTRY_MODE_DATA for ProxyComboEntry (which has a
treeview) and that broke _update_current_object for ProxyEntry using
ENTRY_MODE_DATA.

Do it right by putting the combo specific code on comboentry module. entry's
_update_current_object is called on text change and after that, it emits
changed event that will call the code on the combo (if the entry is on a combo,
of course).

Change-Id: I63137e848e3e9bc5d6649bde19f8613cfa51a153