Don't allow invalid values on CellRendererSpin
There's something wrong with CellRendererSpin when typing the value directly
from the keyboard.

When choosing it with the mouse, it validates right and make sure the value is
on [lower, upper] range.

But when typing the value (using the keyboard) different from the normal spin
that would set the value to min(lower, value) or max(upper, value) it's not
doing that. Do that validation by hand (so it acts exactly like a spin button
would) but add a FIXME message so we can find a better way to solve it in the
future.

Change-Id: Ia9d6556cd23bd1214caf2ae0f2a7ad7a4158c671