Fix an issue with summary label.
The summary labels uses the method `as_string()` from the Column definition,
but if the column defines a `format_func_data` the as_string was passing the
value to the format_func, instead of passing the expected object, which was
causing an error.

Improve the as_string method to better detect when to call format_func.

Change-Id: I82482200b01896e0ddd1119634502c6d4beed305