Allow to set mode by hand without accessing a private attr
The entry behaves different depending on it's actual mode, but data mode is
only set when it gets prefilled with some data. That's a problem when we are
using it and we would need to workaround every api call if the sequence we
passed to prefill was empty.

So, add a new api that allows us to set the mode by hand.

Change-Id: Iff412cd25b51dd8e46e5b8ad5e1af2ff349d12d3