ProxyEntry: Improve validation
Not only KiwiEntry was calling something from it's subclass, but because of
that the validations was being done twice.

That was kind of proxy widget's fault since it doesn't provide a way for
telling if some value for some widget should be considered invalid, only
conversion problems gets caught.

So, add a new api for doing that validation that by default does nothing, but
subclasses can implement it and raise ValidationError if the value should be
considered invalid. Implement it on ProxyEntry and make ProxyComboEntry call it
too.

Change-Id: Ie22daabc6d6c37c3b5009a0ab39de77bda24d1fd