Avoid widget looking weird when invalid and insensitive
When the widget is insensitive and also invalid, avoid drawing a pixbuf
on it since it will make the widget look weird.

Also, that's the correct behaviour since when the widget isn't sensitive
it's not considered invalid by the view so showing an error when insensitive
is confusing for the user.

Change-Id: Ibc5538680cd4e7f522aa5971493190e1040745b5