ProxyWidget: One more improvement to non-sensitive state
Instead of just returning, make sure we set the pixbuf to None, so if there was
a validation error icon on the widget, it will be removed when the widget gets
insensitive.

Also, connect notify::visible too since sensitive/visible are handled in the
same way, that way we make sure that if the widget goes from an
insensitive/invisible state to a sensitive/visible state it will validate
itself again.

Change-Id: I6262d90266db91f2e6d233345b9e002653af13a0