Improve entry completion performance
The match function used needs to strip accents from the string and convert
it to lower case, but this operations are quite time consuming when working
with a lot of objects in the model.

This patch caches the results of this transformation after the first time it
is done. It doesnt make much difference the first time the completion is
executed, but subsequent executions are improved considerably

Change-Id: Idd2698b296b8f9de6e87fffc8bcb4b2c65b8d9d1