Get validation values from the widget
Instead of duplicating the code for widget validity, ask it directly. When
creating a Proxy, all widgets will have it's validate method called, so we can
use widget.is_valid() to check that.

This is more reliable and fix a case where a mandatory empty entry with mask
would be considered valid after adding a proxy. That because the view was
considering "empty" just None and ValueUnset, while the widget.validate()
considers an empty string too.

Change-Id: Ie7e7ec17ce6e97cc91825babd5b12d79410a0a9e