Fix gtkbuilder textdomain translation issue
It turns out that the problem was not that we were not setting textdomain,
but that we calling set_translation_domain too late. The translations seems
to happen when calling "builder.add_from_file"/"builder.add_from_string".

This should work now as long as the textdomain is binded in both gettext
and locale.

Change-Id: I9f314c7c3653ddc8c02079152618efbe0657260c