Workaround issue with gtk3 translation on windows.
Windows with python3 and Gtk3/pygi has this really nasty bug that translations
are actually working, but somehow the enconding is messed up by Gtk. This might
be fixed in newer versions of Gtk/pygi, but currently, we are stuck with 3.24
on Windows.

The idea of the workaround here is to parse the glade xml, and translate in
python all translatable strings, dump the xml on a temp file and then
load the interface from this tempfile. Works like a charm

Other references from a user with the same issue
https://stackoverflow.com/questions/32037573/
https://sourceforge.net/p/pygobjectwin32/tickets/22/
https://bugzilla.gnome.org/show_bug.cgi?id=753991

And the source of the workaround

https://github.com/tobias47n9e/pygobject-locale/issues/1#issuecomment-222287650

Change-Id: I7f381e4de0ad9796e27d4c8e1c2c9e2331f25e91